Refactor internal interfaces together
