+++ internal/contract/interfaces.go
@@ -1,4 +1,4 @@
-// Package contract provides interfaces and shared utilities for the Hotspot CLI's internal architecture.
+// Package contract provides interfaces and shared utilities for internal architecture.
 package contract
 
 import (
@@ -42,4 +42,18 @@ type GitClient interface {
 
 	// GetOldestCommitDateForPath retrieves the commit date of the Nth oldest commit for a path.
 	GetOldestCommitDateForPath(ctx context.Context, repoPath string, path string, before time.Time, numCommits int, maxSearchDuration time.Duration) (time.Time, error)
+}
+
+// CacheManager defines the interface for managing cache stores.
+// This allows the cache layer to be mocked for testing.
+type CacheManager interface {
+	GetActivityStore() CacheStore
+}
+
+// CacheStore defines the interface for cache data storage.
+// This allows mocking the store for testing.
+type CacheStore interface {
+	Get(key string) ([]byte, int, int64, error)
+	Set(key string, value []byte, version int, timestamp int64) error
+	Close() error
 }
